Simplifying the Mediator::GetNodeMode
This function can only tell whether the peer is lost due to winning PoW1. This function is called during ProcessSetShardingInfo in Lookup. The DSCommitteeInfo is obtained earlier than the sharding information, so to be simple can just check whether this peer is happened to be the first element in m_DSCommitteeInfo.