Moved sharing assignments from FINALBLOCK to SHARDING
Now that one DS epoch can have multiple TX epochs, we don't want the
DS nodes to have to keep sending the same txn sharing assignments to
the shard nodes, which is currently the case because the assignments
are appended to the FINALBLOCK message. Now, these assignments have
been moved to the SHARDING message, which is only sent once after
every new DS epoch.