Determine shard membership equivalently for CONTRACT_CALL and CONTRACT_CREATE.
Unlike Scilla, creating a contract involves running the EVM code (constructor).
As we only run EVM on DS at the moment, we must use the same logic for both creation
and calls.

This change also does the same for Scilla, which should not be a problem as contract
creation is rare.