Started with the live 8.1.2 version of production code for AccountStoreSC.tpp and added an intercept to the transaction if it is destined for EVM enabled nodes.
This allows easy comparison of the scilla code path that is in-production, only two minor changes detected with the ScillaIpcServer Inititialisation.

This Commit uses the current EVM enabled UpdateAccounts routine renamed as UpdateAccountEvm with the scilla processing included.

A Subsequent PR will remove the Scilla Code out of EVM code path.

This is for internal testing Only. Has passed all tests with evm_ds_test.