Gas price fixes.
The logic of this gas price is the following:

Min gas price of 0.002 ZIL (2,000,000 gwei) divided by 420 (21000 / 50)
would give 4761. However the price in the testnet is closer to 2x the min
price.

100000 ensures that the price is about 2x the rounded up 5000 gwei price.

Once the 420 factor fix is in, we can remove the 420 multiplier, and leave just 10000 gwei as the gas price.