Revert "Protect m_txnProcessingFinished with a mutex consistently so that (#3502)" (#3581)
This reverts commit 2a7f4f371cc793c51a2bf6675dd987415fdf0a89.

The bug we found in this commit is that one of the `notify_all` wake
up points does not set the `m_txnProcessingFinished` flag, but is
still meant to notify the waiting thread to progress. With the new
overload of `wait_for`, the call is not returned from unless the
condition is met, despite the `notify_all` call. This means the
sleeping thread does not wake, when it previously did.

This is clearly a very poor use of condition variables - We're
using the same condition variable in two different code paths in
two different ways - The first uses it to synchronise access to
the `m_txnProcessingFinished` flag, while the second uses it to
synchronise execution points like a barrier.

We could re-attempt to fix this, however we feel that the chances
of introducing another subtle bug are quite high, so it is safer
to revert the refactor for now. The impact is that we
re-introduce the race condition that the refactor was fixing.
However, the race condition was quite rare and hasn't been
observed in the current mainnet. When the race condition does
occur, it should occur independently in each node and will just
cause those nodes to go out of sync temporarily until they are
restarted.