Add `ots_getContractCreator`
To do this, we add a new `contractCreators` database which stores
a map of contract addresses to the transaction hash in which that
contract was created. This allows us to look up the creator of the
contract in most cases by using the `from` address of the creation
transaction.

Note that this doesn't work for contracts which have been deployed
by other contracts. The API documentation specifically mentions we
should support this, so I've left a `FIXME` for now. To do this,
we'll probably want to extend the data stored when we create a
contract to also track the immediate caller (and thus the contract
creator).