ZIL-5278: Retry sending outgoing messages a fixed number of times
We keep the existing time-based retries, but increase the default
value from 5 seconds to 300 seconds. My intention here is to
effectively disable the time-based timeouts, but keep the option
to re-enable it with a configuration change.

We also start respecting the `MAXRETRYCONN` configuration again
which limits the number of times we will retry the sending of a
message.