[q4-working-branch] Q4 network plus (#3743)
* Add TODOs to fix

* p2pcomm refactor pt.1

* update #2, sending part, relaxed blacklisting policy

* update #3, some wiring

* fixing one stupid thing

* more fixes

* redundant log message eliminated

* dont assigh empty pubkeys

* fix #1

* hotfix about l2 comm

* extra logging

* fix cv_setRejoinRecovery

* fix cv_dsInfoUpdate

* fix cv_shardStruct

* fix cv_vcDsBlockProcessed

* fix cv_vcFinalBlockProcessed

* fix cv_dsLeaderTxnPool

* fix cv_setStateDeltasFromSeed

* fix cv_setCosigRewardsFromSeed

* p2p distributed test tool

* build fix

* extra logging

* Fix Dockerfile to add neccassary apt-get update

* outgoing connections to multiplier nodes follow the old protocol (w/reconnects)

* attempt to fix GetDSLeaderTxPool

* libevent gone

* some TODOs explained

* Just a checkpoint locally

* Fixing test

* Made modifications to logger to make LOG_MARKER very quiet
Changed flag in lookup.cpp to send everything to dsgaurd
Replaced isolated-server-accounts.json with single entry for use with the scripts :
   FundFromPrimaryAccount.ts in the hardhat EvmAcceptanceTests script dir
Updated constants.xml
   open telemetry settings for localdev.py
   reduced extra time for consensus as described in Zelensky doc.

Did basic startup tests

Note this is derived from master

updated localdev.py to start the appropriate number of nodes

localdev.py now defaults to using isolated-server-accounts

./scripts/localdev.py up

* Made modifications to logger to make LOG_MARKER very quiet
Changed flag in lookup.cpp to send everything to dsgaurd
Replaced isolated-server-accounts.json with single entry for use with the scripts :
   FundFromPrimaryAccount.ts in the hardhat EvmAcceptanceTests script dir
Updated constants.xml
   open telemetry settings for localdev.py
   reduced extra time for consensus as described in Zelensky doc.

Did basic startup tests

Note this is derived from master

updated localdev.py to start the appropriate number of nodes

localdev.py now defaults to using isolated-server-accounts

./scripts/localdev.py up

* Increasing POW window

* Upgraded Blacklist to to be keyed by ip/port/node combo.
This is only a partial implementation, need to do the same for whitelist and push through rest of arch.

This change needs to be done in such a way that preserves existing functionality at the ip level.

* Upgraded Blacklist to to be keyed by ip/port/node combo.
This is only a partial implementation, need to do the same for whitelist and push through rest of arch.

This change needs to be done in such a way that preserves existing functionality at the ip level.

* Upgraded Blacklist to to be keyed by ip/port/node combo.
This is only a partial implementation, need to do the same for whitelist and push through rest of arch.

This change needs to be done in such a way that preserves existing functionality at the ip level.

* Upgraded Blacklist to to be keyed by ip/port/node combo.
This is only a partial implementation, need to do the same for whitelist and push through rest of arch.

This change needs to be done in such a way that preserves existing functionality at the ip level.

* Fixing errorneous test

* Upgraded Blacklist to to be keyed by ip/port/node combo.
This is only a partial implementation, need to do the same for whitelist and push through rest of arch.

This change needs to be done in such a way that preserves existing functionality at the ip level.

* Reverted newer tests

* Reverted newer tests
Turned off special handling for multiplier, will remove code on next pass.

* Reverted newer tests
Turned off special handling for multiplier, will remove code on next pass.

* Some in-progress work

* Change genesis account

* Reverted newer tests
Turned off special handling for multiplier, will remove code on next pass.

---------

Co-authored-by: Valeriy Zamaraiev <valeriy@zilliqa.com>
Co-authored-by: Artem <artgor4github@gmail.com>
Co-authored-by: bzawisto <bartosz@zilliqa.com>