ZIL-5376: Support type 1 and type 2 Ethereum transactions (#3827)
A brief description of how these transaction types differ and
the potentially controversial choices I made while supporting
them...

Type 1 transactions are defined in EIP-2930 and allow access lists
to be provided with transactions. In Zilliqa, access lists are
completely ignored and do not have any affect on gas accounting,
but we do keep track of them so the transaction can be
reconstructed later.

Type 2 transactions are defined in EIP-1559 and accept access
lists, plus they allow for a different model for gas accounting.
Instead of setting a single `gasPrice` for a transaction, you set
a 'priority fee' which is paid in addition to the block's base
fee. A maximum fee is also sent in transactions which limits the
total fee the user is willing to pay (base fee + priority fee).
In Zilliqa, we ignore the priority fee and just use the maximum
fee as the entire fee for the transaction.

Its also worth noting that I removed the ability to
configurability of transaction versions and just hardcoded the
values that we use everywhere. We were already hardcoding these
in many places in the source code, so there was only one set of
configured values which would actually result in a working network
anyway.