ZIL-5496: Don't validate work header if extra data was submitted
There are two code paths into `VerifySubmit`. One comes from the
`eth_submitWork` API and does not pass any `extraData`. In this
case we retain the old behaviour and still validate the submitted
header is the same as the header we expect for this bit of work.
The other case comes from the `zil_submitWorkWithExtraData` API.
This takes `extraData` instead of a `header` and computes a new
header based on the data. In this case, we skip the check which
validates the calculated header is the same as the original work
header, because it will always fail.